+++ @navikt/aksel-icons/.svgrrc.js
@@ -1,0 +1,16 @@
+const template = require("./config/template");
+
+module.exports = {
+  typescript: true,
+  ref: true,
+  icon: true,
+  titleProp: true,
+  svgProps: {
+    focusable: false,
+    role: "img",
+  },
+  replaceAttrValues: {
+    "#262626": "currentColor",
+  },
+  template,
+};