+++ aksel.nav.no/website/components/layout/header/Hamburger.tsx
@@ -46,11 +46,11 @@ export const Hamburger = () => {
 
   return (
     <>
-      <div className="z-[1050] mr-0 flex h-full justify-center">
+      <div className="z-[1050] mr-0 grid h-full place-content-center">
         <button
           onClick={() => setOpen(!open)}
           aria-expanded={open}
-          className="focus-visible:shadow-focus hover:bg-surface-neutral-subtle-hover ml-2 flex aspect-square items-center justify-center rounded px-2 py-2 focus:outline-none"
+          className="focus-visible:shadow-focus hover:bg-surface-neutral-subtle-hover ml-2 flex aspect-square items-center justify-center rounded px-3 py-3 focus:outline-none"
         >
           <HamburgerIcon
             className="pointer-events-none text-2xl"
@@ -63,7 +63,7 @@ export const Hamburger = () => {
           aria={{ modal: true }}
           overlayClassName="header-modal__overlay"
           contentLabel="Meny"
-          className="bg-surface-default absolute left-4 top-0 right-4 block rounded py-14 px-11 sm:right-6 sm:left-auto sm:w-96 sm:max-w-[90%] md:hidden"
+          className="bg-surface-default absolute left-4 top-0 right-4 block rounded py-14 px-11 sm:right-6 sm:left-auto sm:w-96 sm:max-w-[90%] lg:hidden"
         >
           <nav aria-label="hovedmeny">
             <ul>
@@ -82,6 +82,11 @@ export const Hamburger = () => {
                 onClick={() => setOpen(false)}
                 name="Grunnleggende"
                 href="/grunnleggende"
+              />
+              <LinkElement
+                onClick={() => setOpen(false)}
+                name="Ikoner"
+                href="/ikoner"
               />
               <LinkElement
                 onClick={() => setOpen(false)}