+++ aksel.nav.no/website/components/website-modules/suggestionblock/index.tsx
@@ -36,6 +36,34 @@ const SuggestionBlockComponent = ({
   );
 };
 
+const SuggestionBlockIcon = ({
+  variant,
+  reference = "",
+}: {
+  variant: "ikon" | "ikon-ny";
+  reference?: string;
+}) => {
+  return (
+    <div className={cl("mb-12 mt-6 flex gap-3 rounded-lg  py-4 ")}>
+      <span className="shrink-0 text-2xl">{options[variant]?.icon}</span>
+      <div className="grid">
+        <Heading size="small" level="2">
+          {options[variant]?.heading}
+        </Heading>
+        <BodyLong className="mt-2">{options[variant]?.text}</BodyLong>
+        <Button
+          variant="secondary"
+          as="a"
+          href={options[variant].link + reference}
+          className="mt-4 w-fit"
+        >
+          Send innspill
+        </Button>
+      </div>
+    </div>
+  );
+};
+
 export const SuggestionBlock = ({
   variant,
   reference = "",
@@ -47,6 +75,7 @@ export const SuggestionBlock = ({
     | "komponent-ny"
     | "komponent-beta"
     | "ikon-ny"
+    | "ikon-not-found"
     | "ikon";
   reference?: string;
 }) => {
@@ -54,14 +83,18 @@ export const SuggestionBlock = ({
     return <SuggestionBlockComponent variant={variant} reference={reference} />;
   }
 
+  if (variant === "ikon-ny" || variant === "ikon") {
+    return <SuggestionBlockIcon variant={variant} reference={reference} />;
+  }
+
   return (
     <div
       className={cl(
-        "ring-border-subtle flex justify-between gap-3 rounded-lg px-6 py-4 ring-1",
+        "ring-border-subtle flex justify-between gap-3 rounded-lg ",
         {
-          "mb-0 mt-5": variant === "ikon" || variant === "ikon-ny",
-          "mb-12": variant !== "ikon" && variant !== "ikon-ny",
-          "bg-surface-info-subtle": variant === "ikon-ny",
+          "mb-12 px-6 py-4 ring-1": variant !== "ikon-not-found",
+          "m-0 bg-teal-100 px-6 py-4 ring-1 ring-teal-300":
+            variant === "ikon-not-found",
         }
       )}
     >
@@ -118,6 +151,12 @@ const options: {
     link:
       issueUrl +
       "&labels=forespørsel+🥰%2Cnytt+✨%2Cikoner+🖼&template=update-icon.yml&title=%5BTilbakemelding%20p%C3%A5%20ikon%5D%3A+",
+  },
+  "ikon-not-found": {
+    text: "Har du forslag til nye ikoner? Trykk på bidra eller send inn et nytt forslag!",
+    link:
+      issueUrl +
+      "&labels=nytt+✨%2Cikoner+🖼%2Cforespørsel+🥰&template&template=new-icon.yaml&title=%5BNytt+ikon%5D%3A+",
   },
   "komponent-ny": {
     text: "Denne komponenten er ny eller oppdatert. Tar du den i bruk ønsker vi gjerne innspill til hvordan den fungerer i tjenesten din!",